AP-18690: Disable dictionary encoding to circumvent a Bug in our Python Ext Types
It looks as if our Python dictionary encoding mechanism does not work as expected
if it is inside of a struct. We did not obtain the decoded value of the CRS, but
always a tuple of (index, optional dictionary entry) which we cannot decode
into a GeoValue without knowing the full column. We created a separate bug ticket
for that and disabled dictionary encoding instead for the GeoValues for the time being.

AP-18690 (GeoSpatial Python Extension Type)